requirements.txt
functions-framework==3.*
google-genai>=1.0.0
google-cloud-storage>=2.0.0
python-dotenv>=1.0.0
flask>=2.0.0
